Вынесла в отдельный класс регистрацию Update связанных классов. ПРоверила- приложение запускается
