+++ src/main/java/doctorBookingApp/controller/UserController.java
@@ -1,11 +1,12 @@
  package doctorBookingApp.controller;
 
- import doctorBookingApp.dto.NewUserDTO;
+ import doctorBookingApp.dto.usersDTO.NewUserDTO;
  import doctorBookingApp.dto.StandardResponseDto;
- import doctorBookingApp.dto.UserDTO;
+ import doctorBookingApp.dto.usersDTO.UserDTO;
  import doctorBookingApp.exeption.RestException;
- import doctorBookingApp.service.ConfirmationCodeService;
- import doctorBookingApp.service.UserService;
+ import doctorBookingApp.service.userService.ConfirmationCodeService;
+ import doctorBookingApp.service.userService.RegistrationUserService;
+ import doctorBookingApp.service.userService.UserService;
  import io.swagger.v3.oas.annotations.Operation;
  import io.swagger.v3.oas.annotations.media.Content;
  import io.swagger.v3.oas.annotations.media.Schema;
@@ -23,31 +24,48 @@
  public class UserController {
 
      private final UserService userService;
-     private final ConfirmationCodeService confirmationCodeService;
+     private final RegistrationUserService registrationUserService;
 
 
      //РЕГИСТРАЦИЯ ПОЛЬЗОВАТЕЛЯ
 
+     @Operation(summary = "Регистрация нового пользователя")
+     @ApiResponses(value = {
+             @ApiResponse(responseCode = "200", description = "Регистрация практически завершена, проверьте свою электронную почту.",
+                     content = @Content(mediaType = "application/json")),
+             @ApiResponse(responseCode = "409", description = "Пользователь с таким email уже существует",
+                     content = @Content(mediaType = "application/json"))
+     })
 
      @PostMapping("/register")
-     public ResponseEntity<String> registerUser(@RequestBody NewUserDTO newUser) throws MessagingException, RestException {
-         userService.registrationUser(newUser);
-         return ResponseEntity.ok("Регистрация практически закончена. Проверьте свой адрес электронной почты на наличие кода подтверждения");
+     public ResponseEntity<String> registerUser(@RequestBody NewUserDTO newUser) throws RestException, MessagingException {
+         registrationUserService.registrationUser(newUser);
+         return ResponseEntity.ok("Регистрация практически завершена. Проверьте свой электронный почтовый ящик на наличие кода подтверждения.");
      }
+
+
+
+     @Operation(summary = "Регистрация. Подтверждение пользователя по конфирмационному коду")
+     @ApiResponses(value = {
+             @ApiResponse(responseCode = "200", description = "Пользователь успешно подтвержден",
+                     content = @Content(mediaType = "application/json")),
+             @ApiResponse(responseCode = "400", description = "Код не найден или срок его действия истек",
+                     content = @Content(mediaType = "application/json"))
+     })
 
 
      @PostMapping("/confirm")
      public ResponseEntity<?> confirmUser(@RequestParam String confirmationCode) {
          try {
-             UserDTO userDTO = confirmationCodeService.confirmation(confirmationCode);
-             return ResponseEntity.ok(userDTO);
+             registrationUserService.confirmUser(confirmationCode);
+             return ResponseEntity.ok("Пользователь успешно подтвержден");
          } catch (RestException e) {
-             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Неверный код подтверждения");
+             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Код не найден или срок его действия истек");
          }
      }
 
 
-     //ПОЛУЧЕНИЕ ИНФОРМАЦИИ О ПОЛЬЗОВАТЕЛЕ
+     //ПОЛУЧЕНИЕ/ОБНОВЛЕНИЕ ИНФОРМАЦИИ О ПОЛЬЗОВАТЕЛЕ
 
      @Operation(summary = "Получение информации о пользователе по ID")
      @ApiResponses(value = {
@@ -77,7 +95,6 @@
                              schema = @Schema(implementation = StandardResponseDto.class)))
 
      })
-
 
      @PutMapping("/{userId}")
      public ResponseEntity<UserDTO> updateUser(@PathVariable Long userId, @RequestBody UserDTO userDTO) throws RestException {
@@ -151,11 +168,6 @@
          userService.deleteUserByPhoneNumber(phoneNumber);
          return ResponseEntity.noContent().build();
      }
-
-
-
-
-
