+++ src/main/java/doctorBookingApp/dto/usersDTO/NewUserDTO.java
@@ -1,4 +1,4 @@
-package doctorBookingApp.dto;
+package doctorBookingApp.dto.usersDTO;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Email;
