+++ src/main/java/doctorBookingApp/repository/UserRepository.java
@@ -1,11 +1,10 @@
 package doctorBookingApp.repository;
 
-import doctorBookingApp.dto.UserDTO;
+
 import doctorBookingApp.entity.ConfirmationCode;
 import doctorBookingApp.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
 import java.util.Optional;
 
 
@@ -28,10 +27,6 @@ public interface UserRepository extends JpaRepository<User, Long>{
 
     Optional<User> findFirstByCodesContains(ConfirmationCode confirmationCode);
 
-//    List<User> findByFullName(String surName, String firstName);
-
-
-//    List<User> findAll (Long id);
